+++ skbio/stats/tests/test_misc.py
@@ -12,6 +12,7 @@ from unittest import TestCase, main
 import numpy as np
 
 from skbio.stats import p_value_to_str
+from skbio.stats._misc import _pprint_strs
 
 
 class PValueToStrTests(TestCase):
@@ -50,6 +51,48 @@ class PValueToStrTests(TestCase):
 
         obs = p_value_to_str(np.nan, 0)
         self.assertEqual(obs, 'N/A')
+
+
+class PPrintStrsTests(TestCase):
+    def test_truncation(self):
+        # truncation between items (on comma)
+        exp = "'a', ..."
+        obs = _pprint_strs(['a', 'b', 'c'], max_chars=4)
+        self.assertEqual(obs, exp)
+
+        # truncation between items (on space)
+        exp = "'a', ..."
+        obs = _pprint_strs(['a', 'b', 'c'], max_chars=5)
+        self.assertEqual(obs, exp)
+
+        # truncation on item
+        exp = "'a', ..."
+        obs = _pprint_strs(['a', 'b', 'c'], max_chars=6)
+        self.assertEqual(obs, exp)
+
+        # truncation (no items)
+        exp = "..."
+        obs = _pprint_strs(['a', 'b', 'c'], max_chars=2)
+        self.assertEqual(obs, exp)
+
+    def test_no_truncation(self):
+        exp = "'a'"
+        obs = _pprint_strs(['a'], max_chars=3)
+        self.assertEqual(obs, exp)
+
+        exp = "'a', 'b', 'c'"
+        obs = _pprint_strs(['a', 'b', 'c'])
+        self.assertEqual(obs, exp)
+
+        exp = "'a', 'b', 'c'"
+        obs = _pprint_strs(['a', 'b', 'c'], max_chars=13)
+        self.assertEqual(obs, exp)
+
+    def test_non_default_delimiter_and_suffix(self):
+        exp = "'abc','defg',...."
+        obs = _pprint_strs(['abc', 'defg', 'hi', 'jklmno'], max_chars=14,
+                           delimiter=',', suffix='....')
+        self.assertEqual(obs, exp)
 
 
 if __name__ == '__main__':