+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 ## Version 0.2.1-dev (changes since 0.2.1 release go here)
 
+### Features
+* Added ``plot`` method to ``skbio.stats.distance.DissimilarityMatrix`` for creating basic heatmaps of a dissimilarity/distance matrix (see [#684](https://github.com/biocore/scikit-bio/issues/684)). Also added  ``_repr_png_`` and ``_repr_svg_`` methods for automatic display in the IPython Notebook, with ``png`` and ``svg`` properties for direct access.
+* Added `__str__` method to `skbio.stats.ordination.OrdinationResults`.
+
+### Bug fixes
+* Fixed issue where SSW wouldn't compile on i686 architectures ([#409](https://github.com/biocore/scikit-bio/issues/409)).
+
 ## Version 0.2.1 (2014-10-27)
 
 This is an alpha release of scikit-bio. At this stage, major backwards-incompatible API changes can and will happen. Unified I/O with the scikit-bio I/O registry was the focus of this release.